Message de validation et d'erreur lors de la suppression d'un commentaire

+++ script/script_delete_commentaire.php
@@ -15,14 +15,29 @@ if(!preg_match('#^[0-9]$#', $idCom) && !is_numeric($idUti)) {
 
 $intIdCom = intval($idCom);
 
-$queryDeleteCom = 'DELETE FROM Commentaire WHERE Id_Utilisateur = :idUti AND Id_Commentaire = :idCom';
-$deleteCom = $mysqlClient->prepare($queryDeleteCom);
-$deleteCom->execute([
-    'idUti' => $idUti,
-    'idCom' => $intIdCom,
+// Selection du blog correspondant à l'id reçus
+$selectBlog = $mysqlClient->prepare('SELECT Id_Blog FROM Blog WHERE Id_Blog = :idBlog');
+$selectBlog->execute([
+    'idBlog' => $idBlog,
 ]);
+$verifBlog = $selectBlog->fetch();
 
-$_SESSION['ERROR_MESSAGE_DELETE_COMMENTAIRE'] = "";
+if($verifBlog == NULL) {
+    $_SESSION['ERROR_MESSAGE_DELETE_COMMENTAIRE'] = "Le commentaire n'existe pas ou vous n'êtes pas le créateur";
 
-header("Location: /../mon_compte.php?page=mes_commentaires");
-exit();
+    header("Location: /../mon_compte.php?page=mes_commentaires");
+    exit();
+} else {
+    // Sureppresion du commentaire correspondant à l'id reçus
+    $queryDeleteCom = 'DELETE FROM Commentaire WHERE Id_Utilisateur = :idUti AND Id_Commentaire = :idCom';
+    $deleteCom = $mysqlClient->prepare($queryDeleteCom);
+    $deleteCom->execute([
+        'idUti' => $idUti,
+        'idCom' => $intIdCom,
+    ]);
+    
+    $_SESSION['VALIDATE_MESSAGE_DELETE_COMMENTAIRE'] = "Votre commentaire a bien été supprimer !";
+    
+    header("Location: /../mon_compte.php?page=mes_commentaires");
+    exit();
+}
